packagegroup-venus-base: include decent utils
The busybox utils used to be installed, but they only implement a
subset of what the normal tools support. That is really annoying at
times, since even the most basic commands no longer work. Since we
will be moving to even far bigger frameworks, which won't fit on the
smaller devices, lets use the available space on those to have
decent tools. This adds a couple of megabytes, the bigger ones count
in gigabytes.